⚡ Bolt: [performance improvement] Optimize graph summary calculation in AdvancedMissionBuilder
💡 What: Replaced the $O(N^2)$ nested `.filter()` and `.some()` logic used to calculate `terminalCount` and `highFanOut` with an $O(N)$ single-pass map-based approach using `fanOutMap`. Also extracted utility functions from `AdvancedMissionBuilder.tsx` to `advancedMissionBuilderUtils.tsx` to pass the <2000 lines CI check.
🎯 Why: Iterating over graph dependencies within another iteration caused significant UI blocking during large complex missions. Also, touching the >2000 line file broke CI.
📊 Impact: Reduced calculation time from quadratic $O(N^2)$ to linear $O(N)$, mitigating freeze risks and dropping unnecessary loop overhead for complex dependencies. Passed the file-size CI limit.
🔬 Measurement: Verify by rendering an Advanced Mission Builder preview with a high volume of node dependencies.
